requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
pydantic-settings==2.7.1
python-dotenv==1.0.1
PyJWT==2.10.1
bcrypt==4.2.1
geopy==2.4.1
python-multipart==0.0.20
pytest==8.3.4
httpx==0.28.1
pytest-cov==6.0.0
gunicorn==23.0.0
